Add `_.iife` for immediately invoking functions.
`_.iife` takes a function as its first argument and calls it immediately,
passing in any additional specified arguments to the function, and then returns
the value of whatever was returned from the called function. Useful as a
semantic wrapper for IIFE's and/or mocking out asynchronous task-based functions
in tests, such as `setTimeout`.