Implement descending order sort in #414.
Change compareMultipleAscending(this, other) to
compareMultiple(this, other, props), which allows sorting in descending
order for a specific property if '-' is prepended to its name.

Change sortByAll so it can sort in descending order for a specific property.
For example, _.sortByAll(users, ['user', '-age']) sort in ascending
order for 'user' but in descending order for 'age'.