Implement optional descending order sort in #414
Change compareMultipleAscending(this, other) to
compareMultiple(this, other, orderArray), which allows sorting descending
order sort for a specific property if its corresponding value in orderArray
is false.

Change sortByAll so it can sort in descending order for a specific
property.
For example, _.sortByAll(users, ['user', { 'key': 'age', 'desc': true }]) sort
in ascending order for 'user' but in descending order for 'age'.