WIP - #1370 Add support for negative indices
This is a POC implementation for supporting negative
indices. There are several performance regressions and shortcuts.

Negative indices are a common pattern that make it easy for developers
to get the last, 2nd-last, value in an array.

The way the implementation was added, it'd have far reaching
implications.

```js
_.result(['a','b','c'], -3) //=> 'a'
_.get(['a','b','c'], -2) //=> 'b'
_.at(['a','b','c'], -1) //=> ['c']
```

It's also possible to support a path push notation where
`a[-0]` could indicate the array location after the last location. This
was brought up by #1315.

```js
_.set({ a: [1] }, 'a[-0]', 2) //=> { a: [1, 2] }
```

@jridgewell and I paired on this.