Add function to make chains callable with a new value
Add _().callable(), which returns a function that executes the actions
of the chain callable() was called on its argument.

This enables the passing of chains to functions which take an interatee.

callable() works by cloning the wrapper and replacing the first __wrapped__
property which is not itself a wrapper with the argument passed to
the function. It then calls value() on the resulting wrapper object
and returns the result.