Add failing test for mergeWith with multi-ref source object
This seems pretty straightfoward to me. I've got a source object
I want to merge which has two keys, only one of which exists in
the object I want to merge into. Both keys in the source object
point to the same object, an array.

What I expect to come out of the merge is an object where the
key that existed before has been modified to contain the values
of the array too, and the key that didn't exist just got added
with the values of the shared object *only*.

What actually happens is the two keys in the final object are
exactly the same, a strict equals comparison is true. This is
unexpected.

Taking this problem a little further, if we create an in initial
object with both keys from the source object defined and set to
different things (e.g. ), then the result of mergeWith will
actually end up destroying data.

The problem appears to stem from `baseMergeDeep`'s stack because
the customiser function is only executed once, for the first key.
It feels like a bad assumption to make is that if you've seen a
source key's value once, the merge result will be the same no
matter what initial key you're merging into.