Run debounce callback if browser tab is closed.
A common issue with debounced API calls is that when the user closes the
browser tab, the trailing debounce may not have run yet. This is a common
cause of data loss. The debounce function prevents this issue by running the
trailing debounce immediately if the tab is closed prior to the trailing
timeout.

The implementation detects whether lodash is running in an environment
with access to `document.addEventListener()`. In runtimes where this
function is unavailable (e.g. Node.js) the new functionality is ignored
since would not be relevant in that context.